[dsd] Add generic helper scripts to fix permissions and remove files. JB#41291
remove-unused.sh is a script which removes unused files as found on different
adaptations.
Files that must be deleted on one adaptation but not on another should go
into copy_system.sh and not in this list.
Files that can be deleted but are only found in one adaptation, should still
be placed here since it doesn't hurt removing something that doesn't exist
in another adaptation and having one complete is better than having multiple
that are out of sync.

fixup-permissions.sh is a script which fixes some permissions before
copying files from system, a similar statement as above applies.

Signed-off-by: Franz-Josef Haider <franz.haider@jolla.com>